Creation time improvements for large clusters when there is no bastion
Run host operations simultaneously. The number of operations to execute
simultaneously is  controlled by a new pnda-env.yaml config
`cli:MAX_SIMULATANEOUS_OUTBOUND_CONNECTIONS`. This setting is defaulted
to 100, if a cluster with more than 100 nodes is being created the
operator can increase this at their discretion.

When there is a bastion, continue to wait 2 seconds between starting
each operation so the bastion does not get overloaded with inbound
connections.

This means the effective wait time when using a bastion still scales
with the number of nodes (number of nodes*4 seconds) and when not using
a bastion it does not increase with the number of nodes assuming
MAX_SIMULATANEOUS_OUTBOUND_CONNECTIONS can be set higher than the
number of nodes in the cluster. If it cannot be set higher then the time will scale with the number of batches the operations must be run in to keep the number of connections below this figure.

PNDA-3553