Add support for configuring Jupyter with SSL cert/key.
PNDA-3623:
The security material should be placed in the
platform-certificates/jupyter/ directory and should comply to the
requirements defined in the platform-certificates/jupyter/README.md
file.
The key will be placed in a pillar accessible to the node with a role
the same name as the subdirectory the key is stored in (Jupyter in this
case).
The certificate will be placed in a shared pillar as that content is not
secret and needs to be accessible to other roles (example: the console
frontend needs to be configured with the CN defined in the certificate
to allow SSL server validation.