Apply timeouts in thread.join
Apply a 10 minute timeout in thread.join rather than relying on an individual thread to timeout and report failure when running commands against minions.

This avoids the case where the CLI calls out to ssh when checking connectivity and it never returns.

This timeout is now applied to all ssh commands on minions (except for the main install of PNDA) and not just the initial connectivity check.

Also added a verbose option -v to the pnda-cli. Setting this sets the logging to debug, defaults to info otherise as is currently the case.

PNDA-4453