Saltmaster tells each minion what its name should be when pre-seeding the keys
Drive the minion ID and names of keys from a single source of data (the names in the instance map) rather than letting each bootstrap script set the minion ID to freeform text.

Now that the saltmaster pre-seeds keys for expected minion IDs it is not sensible to allow free form minion IDs to be set in the bootstrap scripts as if they do not match the ones expected by the saltmaster then the minions will not be allowed to connect.

PNDA-4420